Add files via upload
For any kind of web application login, signup is the most important thing for security reason. If we not have a login features in our web application then any one can access our data and services.
In this project, I have created user registration in PHP with the login script.
On the landing page, it shows a login form with a signup link. The registered user can enter their login details with the login form. Once done, he can get into the dashboard after successful authentication.
If the user does not have an account, then he can click the signup option to create a new account.